+++ src/api/http_request_factory.ts
@@ -67,6 +67,7 @@ export default class HttpRequestFactory {
 		const headers: Record<string, string> = {
 			'app-id': this.appId,
 			'app-secret': this.appSecret,
+			'content-type': 'application/json',
 		};
 		if (this.privateKey) {
 			const publicKey = this.privateKey.to_public_key();